+++ templates/cpp-multi-target/lib/domain/parameters-store.h
@@ -12,13 +12,13 @@ public:
   VectorParametersStore(size_t parameterCount)
       : parameters(parameterCount, 0.0) {}
 
-  float get(uint32_t id) override {
+  double get(uint32_t id) override {
     if (id < parameters.size()) {
       return parameters[id];
     }
-    return 0.f;
+    return 0.0;
   }
-  void set(uint32_t id, float value) override {
+  void set(uint32_t id, double value) override {
     if (id < parameters.size()) {
       parameters[id] = value;
     }
@@ -27,17 +27,17 @@ public:
 
 class UnorderedMapParametersStore : public IParametersStore {
 private:
-  std::unordered_map<uint32_t, float> parameters;
+  std::unordered_map<uint32_t, double> parameters;
 
 public:
-  float get(uint32_t id) override {
+  double get(uint32_t id) override {
     auto it = parameters.find(id);
     if (it != parameters.end()) {
       return it->second;
     }
-    return 0.f;
+    return 0.0;
   }
-  void set(uint32_t id, float value) override { parameters[id] = value; }
+  void set(uint32_t id, double value) override { parameters[id] = value; }
 };
 
 class ParametersStore : public IParametersStore {
@@ -57,9 +57,9 @@ public:
     }
   }
 
-  float get(uint32_t id) override { return impl->get(id); }
+  double get(uint32_t id) override { return impl->get(id); }
 
-  void set(uint32_t id, float value) override { impl->set(id, value); }
+  void set(uint32_t id, double value) override { impl->set(id, value); }
 };
 
 } // namespace sonic
